Move list handler and repository to the selector package
